fix(electron-updater): Allow --package-file arg to escape spaces in filenames (#2739)
Right now package file doesn't escape spaces in the filename, which causes a bug where the package file passed to the nsis installer to be incorrect. This means that when the installer runs, it will fail on folders such as 

`C:\Users\first last\appdata\roaming\package-version.7z`

This fixes it by forcing the whole file path to be a single arg.